ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.22
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.6.1
Pillow==12.1.1
pydantic==2.10.6
